+++ rust/src/router/webrtc_transport.rs
@@ -1071,7 +1071,7 @@ impl WebRtcTransport {
 
     /// Set maximum incoming bitrate for media streams sent by the remote endpoint over this
     /// transport.
-    pub async fn set_max_incoming_bitrate(&self, bitrate: u32) -> Result<(), RequestError> {
+    pub async fn set_max_incoming_bitrate(&self, bitrate: u64) -> Result<(), RequestError> {
         debug!("set_max_incoming_bitrate() [bitrate:{}]", bitrate);
 
         self.set_max_incoming_bitrate_impl(bitrate).await
@@ -1079,7 +1079,7 @@ impl WebRtcTransport {
 
     /// Set maximum outgoing bitrate for media streams sent by the remote endpoint over this
     /// transport.
-    pub async fn set_max_outgoing_bitrate(&self, bitrate: u32) -> Result<(), RequestError> {
+    pub async fn set_max_outgoing_bitrate(&self, bitrate: u64) -> Result<(), RequestError> {
         debug!("set_max_outgoing_bitrate() [bitrate:{}]", bitrate);
 
         self.set_max_outgoing_bitrate_impl(bitrate).await
@@ -1087,7 +1087,7 @@ impl WebRtcTransport {
 
     /// Set minimum outgoing bitrate for media streams sent by the remote endpoint over this
     /// transport.
-    pub async fn set_min_outgoing_bitrate(&self, bitrate: u32) -> Result<(), RequestError> {
+    pub async fn set_min_outgoing_bitrate(&self, bitrate: u64) -> Result<(), RequestError> {
         debug!("set_min_outgoing_bitrate() [bitrate:{}]", bitrate);
 
         self.set_min_outgoing_bitrate_impl(bitrate).await
